feat: integrate voice mode into task board
