+++ dashboard/src/app/tasks/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { useCallback, useEffect, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
+import { useRouter, useSearchParams } from "next/navigation";
 import { useSession } from "next-auth/react";
-import { RiAddLine, RiChatHistoryLine, RiCloseLine, RiFilter3Line, RiNotification3Line, RiNotificationOffLine, RiSearchLine, RiSettings3Line } from "@remixicon/react";
+import { RiAddLine, RiChatHistoryLine, RiMicLine, RiCloseLine, RiFilter3Line, RiNotification3Line, RiNotificationOffLine, RiSearchLine, RiSettings3Line } from "@remixicon/react";
 import {
   getPushState,
   isPushConfigured,
@@ -39,6 +39,7 @@ import { QuickAddTask } from "@/components/tasks/QuickAddTask";
 import { TaskDialog } from "@/components/tasks/TaskDialog";
 import { AddChatDialog } from "@/components/tasks/AddChatDialog";
 import { TaskChatDrawer } from "@/components/tasks/TaskChatDrawer";
+import { VoicePanel } from "@/components/tasks/VoicePanel";
 import { BoardSettingsDialog } from "@/components/tasks/BoardSettingsDialog";
 import { InstallHint } from "@/components/tasks/InstallHint";
 import { useIsMobile } from "@/hooks/useIsMobile";
@@ -48,6 +49,7 @@ const POLL_INTERVAL_MS = 5000;
 export default function TasksPage() {
   const { status: sessionStatus } = useSession();
   const router = useRouter();
+  const searchParams = useSearchParams();
   const isMobile = useIsMobile();
   const [board, setBoard] = useState<TasksBoardResponse | null>(null);
   const [error, setError] = useState<string | null>(null);
@@ -55,6 +57,7 @@ export default function TasksPage() {
   const [editingTask, setEditingTask] = useState<Task | null>(null);
   const [newTaskLane, setNewTaskLane] = useState<string | undefined>(undefined);
   const [settingsOpen, setSettingsOpen] = useState(false);
+  const [voiceOpen, setVoiceOpen] = useState(false);
   const [addChatOpen, setAddChatOpen] = useState(false);
   // Desktop: clicking a non-draft card opens its chat in a side drawer so the
   // board keeps its tab. Task is kept on close for the exit animation.
@@ -67,6 +70,10 @@ export default function TasksPage() {
   const [pushState, setPushState] = useState<PushState | null>(null);
   // Pause polling while a mutation is in flight to avoid clobbering optimistic state.
   const busyRef = useRef(false);
+
+  useEffect(() => {
+    if (searchParams.get("voice") === "1") setVoiceOpen(true);
+  }, [searchParams]);
 
   useEffect(() => {
     if (!isPushSupported()) return;
@@ -178,7 +185,8 @@ export default function TasksPage() {
   const activeTagFilterCount = includedTagIds.length + excludedTagIds.length;
 
   return (
-    <div className="flex h-full flex-col space-y-2 p-4 lg:p-6">
+    <div className="relative flex h-full min-h-0 overflow-hidden">
+      <div className="flex min-w-0 flex-1 flex-col space-y-2 overflow-hidden p-4 transition-[width] duration-300 lg:p-6">
       <div className="pwa-header-offset flex items-center justify-between">
         <h1 className="text-lg font-semibold">Tasks</h1>
         <div className="flex items-center gap-1">
@@ -224,6 +232,15 @@ export default function TasksPage() {
               </DropdownMenuContent>
             </DropdownMenu>
           )}
+          <Button
+            variant={voiceOpen ? "secondary" : "outline"}
+            size="sm"
+            className="gap-1.5"
+            onClick={() => setVoiceOpen(true)}
+          >
+            <RiMicLine className={voiceOpen ? "h-4 w-4 text-emerald-500" : "h-4 w-4"} />
+            <span className="hidden sm:inline">Voice</span>
+          </Button>
           <Button size="sm" onClick={() => void openNewTaskDrawer()}>
             <RiAddLine className="h-4 w-4" />
             New task
@@ -382,6 +399,12 @@ export default function TasksPage() {
           } : current);
         }}
       />
+      </div>
+      {voiceOpen && (
+        <aside className="absolute inset-0 z-40 bg-background md:static md:z-auto md:w-[400px] md:shrink-0 md:border-l md:border-border">
+          <VoicePanel onClose={() => setVoiceOpen(false)} onBoardChange={refresh} />
+        </aside>
+      )}
     </div>
   );
 }
