+++ dashboard/src/components/BottomNav.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { RiAddLine, RiCheckboxLine, RiMicLine, RiMoneyDollarCircleLine } from "@remixicon/react";
+import { RiAddLine, RiCheckboxLine, RiMoneyDollarCircleLine } from "@remixicon/react";
 import { useTaskAttention } from "@/lib/TaskAttentionContext";
 import { useKeyboardVisible } from "@/hooks/useKeyboardVisible";
 import { cn } from "@/lib/utils";
@@ -32,12 +32,6 @@ export default function BottomNav() {
       active: pathname.startsWith("/chat"),
     },
     {
-      name: "Voice",
-      href: "/voice",
-      icon: RiMicLine,
-      active: pathname.startsWith("/voice"),
-    },
-    {
       name: "Usage",
       href: "/my-usage",
       icon: RiMoneyDollarCircleLine,
@@ -46,7 +40,7 @@ export default function BottomNav() {
   ];
 
   return (
-    <nav className="md:hidden shrink-0 grid grid-cols-4 border-t border-border bg-background pb-[max(0.25rem,env(safe-area-inset-bottom))]">
+    <nav className="md:hidden shrink-0 grid grid-cols-3 border-t border-border bg-background pb-[max(0.25rem,env(safe-area-inset-bottom))]">
       {tabs.map((tab) => (
         <Link
           key={tab.name}
